trying some gesture drawing stuff from the tutorial, doesn't fix the problem
right now we have a problem where the live stream image data is not returning any recognition result. The result object is there but all the entries are empty. However, when we just feed in one image there is a result. I'm beginning to think the hand recognition for mediapipe may not work...I'm not sure how to fix this issue.