separating the frame intake and processing for each 'mode' into separate interpreters.
The actual model processing will be contained within these interpreters. The logic will be determined by the GestureStateMachine, which should be renamed to like ModeStateHandler or something, whose direction is determined by the GestureInterpreter. But yeah tomorrow I think we finish this project. At least the mode switching. And even then I think we finish the chess analyzer tomorrow as well. Even the llm thing should be easy since i'll just be hosting an llm on a separate server