renamed stuff, but basically finished the GestureInterpreter, sans a few usage cases
Moved the gesture stuff from the main.py over to the GestureInterpreter. Now I need to integrate that in to the ModeOrchestration. I'm not sure if that's going to be in the main.py or it's own class yet, but we'll see. For the most part the Mode Orchestration will probably be the thing dictating what's actually going on in the glasses I think.

Next I think I should test out that the GestureInterpreter is working well, and I need to figure out how it will work with the Orchestrator and espeically how the callback function will get picked up. We'll see.